refactor: modify ControllerName & change package structure
