+++ src/main/java/racingcar/Application.java
@@ -1,0 +1,11 @@
+package racingcar;
+
+import racingcar.controller.MainController;
+
+public class Application {
+
+    public static void main(String[] args) {
+        MainController game = new MainController();
+        game.start();
+    }
+}
